Add Repository implementations default for each entity

+++ InvoiceSystem.Domain/Interfaces/Repositories/ITaxRepository.cs
@@ -3,6 +3,6 @@ using InvoiceSystem.Domain.Entities;
 
 namespace InvoiceSystem.Domain.Interfaces.Repositories;
 
-public interface ITaxRepository : INPRepository<Tax, int>
+public interface ITaxRepository : INPRepository<Tax, short>
 {
 }
